+++ schemas/productoSchema.py
@@ -8,8 +8,8 @@ class ProductoPydantic(BaseModel):
     marca: str = "P&G"
     categoria: str = "Líquidos"
     cantidad: int = 15
-    valorCompra: float = 4200
-    valorVenta: float = 6500
+    valorCompra: int = 4200
+    valorVenta: int = 6500
     unidadMedida: str = "500ML"
     fechaVencimiento: date = date(2023, 7, 19)
 
@@ -17,8 +17,8 @@ class ProductoUpdatePydantic(BaseModel):
     nombre: str = "Varsol"
     marca: str = "P&G"
     cantidad: int = 15
-    valorCompra: float = 4200
-    valorVenta: float = 6500
+    valorCompra: int = 4200
+    valorVenta: int = 6500
     unidadMedida: str = "500ML"
     fechaVencimiento: date = date(2023, 7, 19)
